refactor(test): move some maven tests into utility tests module
