+++ tests/unit/plugins/test_maven_plugin.py
@@ -14,17 +14,9 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import io
-import os
-import xml.etree.ElementTree as ET
-from pathlib import Path
-from textwrap import dedent
-from unittest import mock
-
 import pytest
 from craft_parts import Part, PartInfo, ProjectInfo, errors
 from craft_parts.plugins.maven_plugin import MavenPlugin
-from craft_parts.utils.maven import create_maven_settings
 from pydantic import ValidationError
 
 
@@ -188,111 +180,3 @@ common/craft-parts/reference/plugins/maven_plugin.html'; exit 1;
             *plugin._get_java_post_build_commands(),
         ]
     )
-
-
-def test_settings_no_proxy(part_info, new_dir):
-    properties = MavenPlugin.properties_class.unmarshal({"source": "."})
-    plugin = MavenPlugin(properties=properties, part_info=part_info)
-
-    # The standard Maven plugin with no configuration options should not
-    # produce a settings file if there aren't any proxies configured
-    with mock.patch.dict(os.environ, {}):
-        assert (
-            create_maven_settings(part_info=plugin._part_info, set_mirror=False) is None
-        )
-
-
-@pytest.mark.parametrize(
-    ("protocol", "expected_protocol"),
-    [
-        pytest.param("http_proxy", "http", id="http"),
-        pytest.param("HTTP_PROXY", "http", id="HTTP"),
-        pytest.param("https_proxy", "https", id="https"),
-        pytest.param("HTTPS_PROXY", "https", id="HTTPS"),
-    ],
-)
-@pytest.mark.parametrize(
-    ("no_proxy", "non_proxy_hosts"),
-    [(None, "localhost"), ("foo", "foo"), ("foo,bar", "foo|bar")],
-)
-@pytest.mark.parametrize(
-    ("credentials", "credentials_xml"),
-    [
-        pytest.param(
-            "username:hunter7@",
-            "<username>username</username>\n<password>hunter7</password>\n",
-            id="with-creds",
-        ),
-        pytest.param("", "", id="no-creds"),
-    ],
-)
-def test_settings_proxy(
-    part_info: PartInfo,
-    protocol: str,
-    expected_protocol: str,
-    no_proxy: str | None,
-    non_proxy_hosts: str,
-    credentials: str,
-    credentials_xml: str,
-):
-    properties = MavenPlugin.properties_class.unmarshal({"source": "."})
-    plugin = MavenPlugin(properties=properties, part_info=part_info)
-    settings_path = Path(
-        part_info._dirs.parts_dir / "my-part/build/.parts/.m2/settings.xml"
-    )
-
-    expected_content = dedent(
-        f"""\
-        <settings xmlns="http://maven.apache.org/SETTINGS/1.0.0"
-                  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                  xsi:schemaLocation="http://maven.apache.org/SETTINGS/1.0.0 http://maven.apache.org/xsd/settings-1.0.0.xsd">
-          <interactiveMode>false</interactiveMode>
-          <proxies>
-            <proxy>
-              <id>{expected_protocol}_proxy</id>
-              <protocol>{expected_protocol}</protocol>
-              <host>my-proxy-host</host>
-              <port>3128</port>
-              <nonProxyHosts>{non_proxy_hosts}</nonProxyHosts>
-              {credentials_xml}
-              <active>true</active>
-            </proxy>
-          </proxies>
-        </settings>
-        """
-    )
-
-    env_dict = {
-        protocol: f"http://{credentials}my-proxy-host:3128",
-    }
-    if no_proxy:
-        env_dict["no_proxy"] = no_proxy
-
-    with mock.patch.dict(os.environ, env_dict):
-        assert plugin.get_build_commands() == (
-            [
-                f"mvn package -s {str(settings_path.absolute())}",
-                *plugin._get_java_post_build_commands(),
-            ]
-        )
-        assert settings_path.exists()
-        assert _normalize_settings(settings_path.read_text()) == _normalize_settings(
-            expected_content
-        )
-
-
-def _normalize_settings(settings):
-    with io.StringIO(settings) as f:
-        tree = ET.parse(f)  # noqa: S314
-    for element in tree.iter():
-        if element.text is not None and element.text.isspace():
-            element.text = None
-        if element.tail is not None and element.tail.isspace():
-            element.tail = None
-    with io.StringIO() as f:
-        tree.write(
-            f,
-            encoding="unicode",
-            default_namespace="http://maven.apache.org/SETTINGS/1.0.0",
-        )
-        return f.getvalue() + "\n"
